Code Review-richrn
Your home.java is extra. Get rid of it. I commented it out, but it doesnt appear to do/effect anything that the calculator doesnt already do.

I added a minimum size frame to the window consider adding dynamic framing.

I added a comment about the odd button in the bottom left. There are some issues with the sizing. As i do not know its purpose i left it as is. Consider moving/removing it.